Refactor: Integer->Long 타입 변경

+++ src/main/java/usedbookshop/soobook/book/book/domain/Book.java
@@ -24,10 +24,10 @@ public class Book extends Date {
     private Long id;
 
     private String title;
-    private Integer price;
-    private Integer score;
+    private Long price;
+    private Long score;
     private String author;
-    private Integer quantity;
+    private Long quantity;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "category_book_id")
@@ -46,7 +46,7 @@ public class Book extends Date {
 
     //private 생성자
     @Builder
-    private Book(String title, Integer price, String author, Integer quantity, Member member) {
+    private Book(String title, Long price, String author, Long quantity, Member member) {
         this.title = title;
         this.price = price;
         this.author = author;
@@ -55,7 +55,7 @@ public class Book extends Date {
         this.member = member;
     }
 
-    public static Book createBook(String title, Integer price, String author, Integer quantity, Member member){
+    public static Book createBook(String title, Long price, String author, Long quantity, Member member){
         return Book.builder()
                 .title(title)
                 .price(price)
@@ -65,7 +65,7 @@ public class Book extends Date {
                 .build();
     }
 
-    public void modifyBook(String title, Integer price, String author, Integer quantity, CategoryBook categoryBook) {
+    public void modifyBook(String title, Long price, String author, Long quantity, CategoryBook categoryBook) {
         this.title = title;
         this.price = price;
         this.author = author;
@@ -82,13 +82,13 @@ public class Book extends Date {
      * 비즈니스 로직
      */
     //quantity 증가
-    public void increaseQuantity(Integer quantity){
+    public void increaseQuantity(Long quantity){
         this.quantity += quantity;
     }
 
     //quantity 감소
-    public void reduceQuantity(Integer quantity){
-        int restStock = this.quantity - quantity;
+    public void reduceQuantity(Long quantity){
+        long restStock = this.quantity - quantity;
         if (restStock < 0) {
             throw new NotEnoughStockException("need more stock");
         }
@@ -102,7 +102,7 @@ public class Book extends Date {
     //책의 평점 계산 후 업데이트
     public void updateScore(){
 
-        int bookScore = 0;
+        long bookScore = 0;
         for (Review review : reviewList){
             bookScore += review.getScore().getValue();
         }
